requirements.txt
Django
djangorestframework
requests
pillow
django-cors-headers
psycopg-binary